# modules/base/docs/base_swagger.yaml
definitions:
  github_com_AgileExecutives_serverbase_internal_models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_AgileExecutives_serverbase_internal_models.ContactFormRequest:
    description: Contact form submission request
    properties:
      email:
        example: john.doe@example.com
        type: string
      message:
        example: I am interested in learning more about your therapy services.
        type: string
      name:
        example: John Doe
        type: string
      newsletter:
        example: true
        type: boolean
      source:
        example: website
        type: string
      subject:
        example: Inquiry about therapy services
        type: string
      timestamp:
        example: "2025-08-03T10:00:00Z"
        type: string
    required:
    - email
    - message
    - name
    - source
    - subject
    type: object
  github_com_AgileExecutives_serverbase_internal_models.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
    type: object
  github_com_AgileExecutives_serverbase_internal_models.ListResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.PaginationResponse'
    type: object
...
  github_com_AgileExecutives_serverbase_internal_models.Newsletter:
    properties:
      createdAt:
        example: "2025-08-03T10:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 1
        type: integer
      interest:
        example: mental_health
        type: string
      lastContact:
        example: "2025-08-03T10:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      source:
        example: website
        type: string
      updatedAt:
        example: "2025-08-03T10:00:00Z"
        type: string
    type: object
  github_com_AgileExecutives_serverbase_internal_models.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.Contact:
    properties:
      active:
        type: boolean
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      mobile:
        type: string
      notes:
        type: string
      phone:
        type: string
      street:
        type: string
      type:
        type: string
      updated_at:
        type: string
      zip:
        type: string
    required:
    - first_name
    - last_name
    type: object
  models.ContactCreateRequest:
    properties:
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile:
        type: string
      notes:
        type: string
      phone:
        type: string
      street:
        type: string
      type:
        type: string
      zip:
        type: string
    required:
    - first_name
    - last_name
    type: object
  models.ContactResponse:
    properties:
      active:
        type: boolean
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      mobile:
        type: string
      notes:
        type: string
      phone:
        type: string
      street:
        type: string
      type:
        type: string
      zip:
        type: string
    type: object
  models.ContactUpdateRequest:
    properties:
      active:
        type: boolean
      city:
        type: string
      country:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile:
        type: string
      notes:
        type: string
      phone:
        type: string
      street:
        type: string
      type:
        type: string
      zip:
        type: string
    type: object
...
info:
  contact: {}
  description: Authentication, contacts, health, and user settings endpoints.
  title: AE SaaS Base API
  version: 1.0.0
paths:
...
  /contacts:
    get:
      consumes:
      - application/json
      description: Get paginated list of contacts with optional filters
      operationId: getContacts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Filter by contact type (business, personal, etc)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ListResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Create a new contact for the authenticated user
      operationId: createContact
      parameters:
      - description: Contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ContactResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create new contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a contact by ID
      operationId: deleteContact
      parameters:
      - description: Contact ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete contact
      tags:
      - contacts
    get:
      consumes:
      - application/json
      description: Get a specific contact by ID
      operationId: getContactById
      parameters:
      - description: Contact ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Contact'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get contact by ID
      tags:
      - contacts
    put:
      consumes:
      - application/json
      description: Update an existing contact by ID
      operationId: updateContact
      parameters:
      - description: Contact ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated contact data
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ContactResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update contact
      tags:
      - contacts
  /contacts/form:
    post:
      consumes:
      - application/json
      description: Public endpoint to submit a contact form (no authentication required)
      operationId: submitContactForm
      parameters:
      - description: Contact form data
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ContactFormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      summary: Submit contact form
      tags:
      - contact-form
  /contacts/newsletter:
    get:
      consumes:
      - application/json
      description: Get all newsletter subscriptions
      operationId: getNewsletterSubscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.Newsletter'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get newsletter subscriptions
      tags:
      - newsletter
  /contacts/newsletter/unsubscribe:
    delete:
      consumes:
      - application/json
      description: Unsubscribe an email address from the newsletter
      operationId: unsubscribeFromNewsletter
      parameters:
      - description: Email address to unsubscribe
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_AgileExecutives_serverbase_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unsubscribe from newsletter
      tags:
      - newsletter
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token – format: ''Bearer <token>'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"